+++ src/models/LineItem.php
@@ -446,7 +446,7 @@ class LineItem extends Model
     {
         if (isset($rule[1]) && $rule[1] instanceof Closure) {
             $method = $rule[1];
-            $method->bindTo($purchasable);
+            $method = $method->bindTo($purchasable);
             $rule[1] = static function($attribute, $params, $validator, $current) use ($method) {
                 $method($attribute, $params, $validator, $current);
             };